VkDestroySwapchainKHR: Account for destroying null chains
mpv, as an example, behaves poorly, and will call vkDestroySwapchainKHR
with VK_NULL_HANDLE as the swapchain upon init. It seems that downstream
drivers tend to handle this gracefully, so we can check to make sure
that the swapchain is either non-null (so we should expect to have a
reference to ours), or we *do* have a reference to a VK_NULL_HANDLE
swapchain (unexpected, but... maybe?) before we dereference it, thereby
allowing the layer to be used with such poorly-behaved applications.